# docs/swagger.yaml
basePath: /
definitions:
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  model.AuthInfo:
    properties:
      authId:
        type: string
      authType:
        type: string
      createAt:
        type: string
      id:
        type: integer
      updateAt:
        type: string
      url:
        type: string
      userId:
        type: integer
    type: object
...
  model.CreatedUser:
    properties:
      authInfos:
        items:
          $ref: '#/definitions/model.AuthInfo'
        type: array
      avatar:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  model.Group:
    properties:
      createAt:
        type: string
      creatorId:
        type: integer
      describe:
        type: string
      id:
        type: integer
      name:
        type: string
      updateAt:
        type: string
      updaterId:
        type: integer
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
...
  model.UpdatedUser:
    properties:
      authInfos:
        items:
          $ref: '#/definitions/model.AuthInfo'
        type: array
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  model.User:
    properties:
      authInfos:
        items:
          $ref: '#/definitions/model.AuthInfo'
        type: array
      avatar:
        type: string
      createAt:
        type: string
      email:
        type: string
      groups:
        items:
          $ref: '#/definitions/model.Group'
        type: array
      id:
        type: integer
      name:
        type: string
      updateAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a weave server api doc.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Weave Server API
  version: "2.0"
paths:
...
  /api/v1/users:
    get:
      description: List user and storage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.User'
                  type: array
              type: object
      security:
      - JWT: []
      summary: List user
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Create user and storage
      parameters:
      - description: user info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.CreatedUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      security:
      - JWT: []
      summary: Create user
      tags:
      - user
  /api/v1/users/{id}:
    delete:
      description: Delete user and storage
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - JWT: []
      summary: Delete user
      tags:
      - user
    get:
      description: Get user and storage
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      security:
      - JWT: []
      summary: Get user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update user and storage
      parameters:
      - description: user info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UpdatedUser'
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      security:
      - JWT: []
      summary: Update user
      tags:
      - user
  /api/v1/users/{id}/groups:
    get:
      description: Get groups
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - JWT: []
      summary: Get groups
      tags:
      - group
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"